- replace Sidebar with layout - added Drawer component
