+++ src/Navigation/Drawer/index.js
@@ -1,0 +1,3 @@
+import Drawer from './Drawer';
+
+export default Drawer;
