+++ src/Navigation/Layout/Layout.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import {
+  LayoutContainer,
+} from '../../Style/Navigation/SidebarStyle';
+
+const Layout = ({ children }:Props) => (
+  <LayoutContainer>
+    {children}
+  </LayoutContainer>
+);
+
+type Props = {
+    children: React$Node,
+}
+
+export default Layout;
