+++ src/Navigation/Layout/LayoutBody.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import {
+  LayoutBody,
+} from '../../Style/Navigation/SidebarStyle';
+
+const LayoutBodyContainer = ({ children, ...defaultProps }:Props) => (
+  <LayoutBody {...defaultProps}>
+    {children}
+  </LayoutBody>
+);
+
+type Props = {
+    children: React$Node,
+}
+
+export default LayoutBodyContainer;
