+++ src/Navigation/Layout/index.js
@@ -1,0 +1,10 @@
+import Layout from './Layout';
+import LayoutHeader from './LayoutHeader';
+import LayoutBody from './LayoutBody';
+import LayoutFooter from './LayoutFooter';
+
+Layout.Header = LayoutHeader;
+Layout.Body = LayoutBody;
+Layout.Footer = LayoutFooter;
+
+export default Layout;
