--- src/Navigation/SideBar/index.js
@@ -1,3 +1,0 @@
-import SideBar from './SideBar';
-
-export default SideBar;
