+++ src/index.js
@@ -14,6 +14,7 @@ export { default as Button } from './General/Button';
 export { default as Checkbox } from './Input/Checkbox';
 export { default as Collapsible } from './Display/Collapsible';
 export { default as Divider } from './General/Divider';
+export { default as Drawer } from './Navigation/Drawer';
 export { default as Dropdown } from './Navigation/Dropdown';
 export { default as Gallery } from './Application/Gallery';
 export { default as Heading } from './General/Heading';
@@ -22,6 +23,7 @@ export { default as Icon } from './General/Icon';
 export { default as JobCard } from './Application/JobCard';
 export { default as JobOverview } from './Application/JobOverview';
 export { default as LanguageSelect } from './Input/LanguageSelect';
+export { default as Layout } from './Navigation/Layout';
 export { default as Loading } from './General/Loading';
 export { default as MobileMenu } from './Navigation/MobileMenu';
 export { default as Modal } from './Display/Modal';
@@ -31,7 +33,6 @@ export { default as Popover } from './Display/Popover';
 export { default as ProfilePicture } from './General/ProfilePicture';
 export { default as PsychedelicText } from './General/PsychedelicText';
 export { default as RadioButton } from './Input/RadioButton';
-export { default as SideBar } from './Navigation/SideBar';
 export { default as SearchFilter } from './Input/SearchFilter';
 export { default as Select } from './Input/Select';
 export { default as Slider } from './Display/Slider';
